Fixed blocky and slow Navigation due to th BuildPage() being called in the OnAppearing, now it is called in the constructor of page. Also made sure animations were fixed.

+++ WindesHeartApp/WindesHeartApp/Views/AboutPage.xaml.cs
@@ -145,7 +145,7 @@ namespace WindesHeartApp.Pages
             Console.WriteLine("Learn More - Clicked.");
             Vibration.Vibrate(4200);
         }
-        public IPageAnimation PageAnimation { get; } = new SlidePageAnimation { Duration = AnimationDuration.Long, Subtype = AnimationSubtype.FromTop };
+        public IPageAnimation PageAnimation { get; } = new SlidePageAnimation { Duration = AnimationDuration.Short, Subtype = AnimationSubtype.FromTop };
 
         public void OnAnimationStarted(bool isPopAnimation)
         {
