+++ WindesHeartApp/WindesHeartApp/Views/DevicePage.xaml.cs
@@ -93,7 +93,7 @@ namespace WindesHeartApp.Pages
             #endregion
 
         }
-        public IPageAnimation PageAnimation { get; } = new SlidePageAnimation { Duration = AnimationDuration.Long, Subtype = AnimationSubtype.FromTop };
+        public IPageAnimation PageAnimation { get; } = new SlidePageAnimation { Duration = AnimationDuration.Short, Subtype = AnimationSubtype.FromTop };
 
         public void OnAnimationStarted(bool isPopAnimation)
         {
