+++ WindesHeartApp/WindesHeartApp/Views/HeartratePage.xaml.cs
@@ -31,7 +31,7 @@ namespace WindesHeartApp.Pages
 
 
         }
-        public IPageAnimation PageAnimation { get; } = new SlidePageAnimation { Duration = AnimationDuration.Long, Subtype = AnimationSubtype.FromTop };
+        public IPageAnimation PageAnimation { get; } = new SlidePageAnimation { Duration = AnimationDuration.Short, Subtype = AnimationSubtype.FromTop };
 
         public void OnAnimationStarted(bool isPopAnimation)
         {
